[DMS-475] EducationOrganizationHierarchy Opensearch query (#433)
* Simplify OpenSearch Query and verify tests still pass

* Tighten up connector. Id field is correct, hierrchy is json

* EdOrg filter query applied. RestClient tests pass

* Back to simplified security element

* Couple of query tests, non hierarchy

* Hierarchy query e2e test

* Support LONG edorg id. Identify bug DMS-556

* Additional test

+++ src/dms/core/EdFi.DataManagementService.Core.External/Model/EducationOrganizationHierarchyInfo.cs
@@ -10,6 +10,6 @@ namespace EdFi.DataManagementService.Core.External.Model;
 /// </summary>
 public record EducationOrganizationHierarchyInfo(
     bool IsInEducationOrganizationHierarchy,
-    int Id,
-    int[] ParentIds
+    long Id,
+    long[] ParentIds
 );
